header only logging.h, add utility.h and utility.cpp
