+++ library/src/blas1/rocblas_scal.cpp
@@ -8,6 +8,8 @@
 
 #include "definitions.h"
 #include "handle.h"
+#include "logging.h"
+#include "utility.h"
 
 #define NB_X 256
