+++ library/src/blas1/rocblas_swap.cpp
@@ -7,6 +7,8 @@
 
 #include "definitions.h"
 #include "handle.h"
+#include "logging.h"
+#include "utility.h"
 
 #define NB_X 256
