+++ library/src/blas3/rocblas_trmm.cpp
@@ -1,11 +1,12 @@
 /* ************************************************************************
  * Copyright 2016 Advanced Micro Devices, Inc.
- *
  * ************************************************************************ */
 #include <hip/hip_runtime.h>
 
 #include "rocblas.h"
 #include "definitions.h"
+#include "logging.h"
+#include "utility.h"
 
 // do not use fma which is 50% slower than regular fmaf
 #define fmaf(a, b, c) (a) * (b) + (c)
