+++ library/src/blas3/rocblas_trsm.cpp
@@ -1,6 +1,5 @@
 /* ************************************************************************
  * Copyright 2016 Advanced Micro Devices, Inc.
- *
  * ************************************************************************ */
 #include <hip/hip_runtime_api.h>
 #include <hip/hip_runtime.h>
